cursed_reformat

+++ src/main/java/growthcraft/milk/common/block/BlockCheesePress.java
@@ -23,14 +23,13 @@
  */
 package growthcraft.milk.common.block;
 
-import growthcraft.api.core.util.BlockFlags;
-import growthcraft.core.common.block.GrcBlockContainer;
-import growthcraft.milk.client.render.RenderCheesePress;
-import growthcraft.milk.common.tileentity.TileEntityCheesePress;
-import growthcraft.milk.GrowthCraftMilk;
-
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import growthcraft.api.core.util.BlockFlags;
+import growthcraft.core.common.block.GrcBlockContainer;
+import growthcraft.milk.GrowthCraftMilk;
+import growthcraft.milk.client.render.RenderCheesePress;
+import growthcraft.milk.common.tileentity.TileEntityCheesePress;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.EntityLivingBase;
@@ -42,174 +41,141 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
 
-public class BlockCheesePress extends GrcBlockContainer
-{
-	public BlockCheesePress()
-	{
-		super(Material.wood);
-		setResistance(5.0F);
-		setHardness(2.0F);
-		setStepSound(soundTypeWood);
-		setBlockName("grcmilk.CheesePress");
-		setCreativeTab(GrowthCraftMilk.creativeTab);
-		setTileEntityType(TileEntityCheesePress.class);
-		setBlockTextureName("grcmilk:cheese_press");
-	}
+public class BlockCheesePress extends GrcBlockContainer {
+    public BlockCheesePress() {
+        super(Material.wood);
+        setResistance(5.0F);
+        setHardness(2.0F);
+        setStepSound(soundTypeWood);
+        setBlockName("grcmilk.CheesePress");
+        setCreativeTab(GrowthCraftMilk.creativeTab);
+        setTileEntityType(TileEntityCheesePress.class);
+        setBlockTextureName("grcmilk:cheese_press");
+    }
 
-	@Override
-	public boolean isRotatable(IBlockAccess world, int x, int y, int z, ForgeDirection side)
-	{
-		return true;
-	}
+    @Override
+    public boolean isRotatable(IBlockAccess world, int x, int y, int z, ForgeDirection side) {
+        return true;
+    }
 
-	public void doRotateBlock(World world, int x, int y, int z, ForgeDirection side)
-	{
-		world.setBlockMetadataWithNotify(x, y, z, world.getBlockMetadata(x, y, z) ^ 1, BlockFlags.SYNC);
-	}
+    public void doRotateBlock(World world, int x, int y, int z, ForgeDirection side) {
+        world.setBlockMetadataWithNotify(x, y, z, world.getBlockMetadata(x, y, z) ^ 1, BlockFlags.SYNC);
+    }
 
-	@Override
-	public void onBlockAdded(World world, int x, int y, int z)
-	{
-		super.onBlockAdded(world, x, y, z);
-		this.setDefaultDirection(world, x, y, z);
-	}
+    @Override
+    public void onBlockAdded(World world, int x, int y, int z) {
+        super.onBlockAdded(world, x, y, z);
+        this.setDefaultDirection(world, x, y, z);
+    }
 
-	private void setDefaultDirection(World world, int x, int y, int z)
-	{
-		if (!world.isRemote)
-		{
-			final Block block = world.getBlock(x, y, z - 1);
-			final Block block1 = world.getBlock(x, y, z + 1);
-			final Block block2 = world.getBlock(x - 1, y, z);
-			final Block block3 = world.getBlock(x + 1, y, z);
-			byte meta = 3;
+    private void setDefaultDirection(World world, int x, int y, int z) {
+        if (!world.isRemote) {
+            final Block block = world.getBlock(x, y, z - 1);
+            final Block block1 = world.getBlock(x, y, z + 1);
+            final Block block2 = world.getBlock(x - 1, y, z);
+            final Block block3 = world.getBlock(x + 1, y, z);
+            byte meta = 3;
 
-			if (block.func_149730_j() && !block1.func_149730_j())
-			{
-				meta = 3;
-			}
+            if (block.func_149730_j() && !block1.func_149730_j()) {
+                meta = 3;
+            }
 
-			if (block1.func_149730_j() && !block.func_149730_j())
-			{
-				meta = 2;
-			}
+            if (block1.func_149730_j() && !block.func_149730_j()) {
+                meta = 2;
+            }
 
-			if (block2.func_149730_j() && !block3.func_149730_j())
-			{
-				meta = 5;
-			}
+            if (block2.func_149730_j() && !block3.func_149730_j()) {
+                meta = 5;
+            }
 
-			if (block3.func_149730_j() && !block2.func_149730_j())
-			{
-				meta = 4;
-			}
+            if (block3.func_149730_j() && !block2.func_149730_j()) {
+                meta = 4;
+            }
 
-			world.setBlockMetadataWithNotify(x, y, z, meta, BlockFlags.UPDATE_AND_SYNC);
-		}
-	}
+            world.setBlockMetadataWithNotify(x, y, z, meta, BlockFlags.UPDATE_AND_SYNC);
+        }
+    }
 
-	@Override
-	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase entity, ItemStack stack)
-	{
-		super.onBlockPlacedBy(world, x, y, z, entity, stack);
-		final int a = MathHelper.floor_double((entity.rotationYaw * 4.0D / 360.0D) + 0.5D) & 3;
-		if (a == 0 || a == 2)
-		{
-			world.setBlockMetadataWithNotify(x, y, z, 0, BlockFlags.SYNC);
-		}
-		else if (a == 1 || a == 3)
-		{
-			world.setBlockMetadataWithNotify(x, y, z, 1, BlockFlags.SYNC);
-		}
-	}
+    @Override
+    public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase entity, ItemStack stack) {
+        super.onBlockPlacedBy(world, x, y, z, entity, stack);
+        final int a = MathHelper.floor_double((entity.rotationYaw * 4.0D / 360.0D) + 0.5D) & 3;
+        if (a == 0 || a == 2) {
+            world.setBlockMetadataWithNotify(x, y, z, 0, BlockFlags.SYNC);
+        } else if (a == 1 || a == 3) {
+            world.setBlockMetadataWithNotify(x, y, z, 1, BlockFlags.SYNC);
+        }
+    }
 
-	@Override
-	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int meta, float par7, float par8, float par9)
-	{
-		if (super.onBlockActivated(world, x, y, z, player, meta, par7, par8, par9)) return true;
-		if (GrowthCraftMilk.getConfig().cheesePressHandOperated)
-		{
-			if (!player.isSneaking())
-			{
-				final TileEntityCheesePress cheesePress = getTileEntity(world, x, y, z);
-				if (cheesePress != null)
-				{
-					if (cheesePress.toggle())
-					{
-						world.playSoundEffect((double)x, (double)y, (double)z, "random.wood_click", 0.3f, 0.5f);
-					}
-					return true;
-				}
-			}
-		}
-		return false;
-	}
+    @Override
+    public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int meta, float par7, float par8, float par9) {
+        if (super.onBlockActivated(world, x, y, z, player, meta, par7, par8, par9)) return true;
+        if (GrowthCraftMilk.getConfig().cheesePressHandOperated) {
+            if (!player.isSneaking()) {
+                final TileEntityCheesePress cheesePress = getTileEntity(world, x, y, z);
+                if (cheesePress != null) {
+                    if (cheesePress.toggle()) {
+                        world.playSoundEffect(x, y, z, "random.wood_click", 0.3f, 0.5f);
+                    }
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
 
-	private void updatePressState(World world, int x, int y, int z)
-	{
-		final boolean isPowered = world.isBlockIndirectlyGettingPowered(x, y, z);
-		final TileEntityCheesePress cheesePress = getTileEntity(world, x, y, z);
-		if (cheesePress != null)
-		{
-			if (cheesePress.toggle(isPowered))
-			{
-				world.playSoundEffect((double)x, (double)y, (double)z, "random.wood_click", 0.3f, 0.5f);
-			}
-		}
-	}
+    private void updatePressState(World world, int x, int y, int z) {
+        final boolean isPowered = world.isBlockIndirectlyGettingPowered(x, y, z);
+        final TileEntityCheesePress cheesePress = getTileEntity(world, x, y, z);
+        if (cheesePress != null) {
+            if (cheesePress.toggle(isPowered)) {
+                world.playSoundEffect(x, y, z, "random.wood_click", 0.3f, 0.5f);
+            }
+        }
+    }
 
-	@Override
-	public void onNeighborBlockChange(World world, int x, int y, int z, Block block)
-	{
-		super.onNeighborBlockChange(world, x, y, z, block);
-		if (!world.isRemote)
-		{
-			if (GrowthCraftMilk.getConfig().cheesePressRedstoneOperated)
-			{
-				this.updatePressState(world, x, y, z);
-			}
-		}
-	}
+    @Override
+    public void onNeighborBlockChange(World world, int x, int y, int z, Block block) {
+        super.onNeighborBlockChange(world, x, y, z, block);
+        if (!world.isRemote) {
+            if (GrowthCraftMilk.getConfig().cheesePressRedstoneOperated) {
+                this.updatePressState(world, x, y, z);
+            }
+        }
+    }
 
-	@Override
-	public int getRenderType()
-	{
-		return RenderCheesePress.RENDER_ID;
-	}
+    @Override
+    public int getRenderType() {
+        return RenderCheesePress.RENDER_ID;
+    }
 
-	@Override
-	public boolean isOpaqueCube()
-	{
-		return false;
-	}
+    @Override
+    public boolean isOpaqueCube() {
+        return false;
+    }
 
-	@Override
-	public boolean renderAsNormalBlock()
-	{
-		return false;
-	}
+    @Override
+    public boolean renderAsNormalBlock() {
+        return false;
+    }
 
-	@Override
-	@SideOnly(Side.CLIENT)
-	public boolean shouldSideBeRendered(IBlockAccess world, int x, int y, int z, int side)
-	{
-		return true;
-	}
+    @Override
+    @SideOnly(Side.CLIENT)
+    public boolean shouldSideBeRendered(IBlockAccess world, int x, int y, int z, int side) {
+        return true;
+    }
 
-	@Override
-	public boolean hasComparatorInputOverride()
-	{
-		return true;
-	}
+    @Override
+    public boolean hasComparatorInputOverride() {
+        return true;
+    }
 
-	@Override
-	public int getComparatorInputOverride(World world, int x, int y, int z, int par5)
-	{
-		final TileEntityCheesePress te = getTileEntity(world, x, y, z);
-		if (te != null)
-		{
-			return Container.calcRedstoneFromInventory(te);
-		}
-		return 0;
-	}
+    @Override
+    public int getComparatorInputOverride(World world, int x, int y, int z, int par5) {
+        final TileEntityCheesePress te = getTileEntity(world, x, y, z);
+        if (te != null) {
+            return Container.calcRedstoneFromInventory(te);
+        }
+        return 0;
+    }
 }
